+++ native/GhostRigger.Tools.Workflow.ModuleMeshes/Python/src/gui/panels/module_editor/__init__.py
@@ -1,4 +1,4 @@
-"""Standalone Module Editor panels."""
+"""Map Studio Level Editor panels."""
 
 from .module_editor_outliner import ModuleEditorOutliner
 from .module_editor_asset_browser import ModuleEditorAssetBrowser
@@ -7,6 +7,7 @@ from .module_editor_toolbar import ModuleEditorToolbar
 from .module_editor_viewport_panel import ModuleEditorViewportPanel
 from .readiness_panel import ModuleReadinessPanel
 from .validation_panel import ModuleValidationPanel
+from .workflow_panel import MapStudioWorkflowPanel
 
 __all__ = [
     "ModuleEditorOutliner",
@@ -16,4 +17,5 @@ __all__ = [
     "ModuleEditorViewportPanel",
     "ModuleReadinessPanel",
     "ModuleValidationPanel",
+    "MapStudioWorkflowPanel",
 ]